Bot_telegram/requirements.txt
<<<<<<< HEAD
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
mysql-connector-python==9.4.0
pyTelegramBotAPI==4.29.1
requests==2.32.5
urllib3==2.5.0
=======
asgiref==3.9.1
Django==5.2.6
sqlparse==0.5.3
tzdata==2025.2
>>>>>>> 47ee0adc3ecaeb76807991d05f40ca0cdde98818
